Apply sign test

This function applies the one-sample sign test to input data grouped by specified variables.

apply_sign_test(
  stats_data,
  summarized_data,
  value = "Value",
  p.signif = "p.signif",
  grouping = NULL,
  na = "NA"
)

Arguments

stats_data

A data frame containing the calculated growth performance data, e.g. via a function call to calculate_growth_performance.

summarized_data

A data frame containing corresponding summarized data, e.g. via function call summarize_growth_performance.

value

The column containing absorption values to be tested. Defaults to 'Value'.

p.signif

The column containing significance denoted in asterisk notation. Defaults to 'p.signif'.

grouping

A character vector specifying the grouping variables.

na

A character value specifying the keyword to display if sign tests cannot be applied on subsets of the data (e.g. because of too small sample sizes). Defaults to "NA".

Value

A data frame containing the summarized data with sign test results added.

See also