Absorption values from six broth microdilution assays conducted on 96-well plates

A list of six sample data sets with absorption values from broth microdilution assays on 96-well plates, applied to two groups with one experiment each at two time points T0 and T3.

bma

Format

A list with six data frames:

bma_grp1_exp2_T0

Absorption values from a broth microdilution assay applied on group 1 from experiment 2 on 96-well plate at timepoint T0.

bma_grp1_exp2_T3

Absorption values from a broth microdilution assay applied on group 1 from experiment 2 on 96-well plate at timepoint T3.

bma_grp2_exp1_T0

Absorption values from a broth microdilution assay applied on group 2 from experiment 1 on 96-well plate at timepoint T0.

bma_grp2_exp1_T3

Absorption values from a broth microdilution assay applied on group 2 from experiment 1 on 96-well plate at timepoint T3.

Details

The data was derived from two broth microdilution assay experiments testing the growth performance of Botrytis cinerea conidia on two Tanacetum vulgare chemotypes (defined as groups). Leaf extracts from chemotypes were fractionated using solid-phase extraction with a water-methanol polarity gradient (defined as treatment) and fractions were subjected to assays in two concentrations (100 ppm and 200 ppm) plus a positive control for each concentration level. The 96-well plate design was assigned in horizontal direction (provided as plate_axis) and is stored in the metadata attribute of the list. The data was generated for teaching purposes and is unrestricted by any licensing constraints.

Examples

data(bma)
attr(bma, "metadata")
#>   plate_axis treatment concentration
#> 1          A       10%       100 ppm
#> 2          B       10%       200 ppm
#> 3          C       30%       100 ppm
#> 4          D       30%       200 ppm
#> 5          E      100%       100 ppm
#> 6          F      100%       200 ppm
#> 7          G   Control       100 ppm
#> 8          H   Control       200 ppm